chore: address clippy lints

+++ crates/phantom_world/src/texture.rs
@@ -118,8 +118,8 @@ impl Texture {
         let decoded = decoder
             .read_image_hdr()
             .map_err(TextureError::DecodeHdrImage)?;
-        let width = metadata.width as u32;
-        let height = metadata.height as u32;
+        let width = metadata.width;
+        let height = metadata.height;
         let data = decoded
             .iter()
             .flat_map(|pixel| vec![pixel[0], pixel[1], pixel[2], 1.0])
